Add files via upload
The first cell of the Jupyter file is the script that generates synthetic microscope images with random patterns representing nanowires. This version also saves the lengths of the nanowires in a CSV file. 

Second cell process the images, detect the nanowires using Hough transform, and compare the detected lengths with the actual lengths.